backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "langchain>=0.3.25",
    "langchain-cohere>=0.4.4",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
]
